Move test keys to a separate source file. Use vector<uint64_t> for s2 since the rest of the system uses this format - it needs to be made consistent with other opaque data types and use vector<uint8_t> eventually. WIP wallet-cli

+++ utt/utt-common-api/include/utt-common-api/CommonApi.hpp
@@ -34,7 +34,8 @@ using UserRegistrationInput = std::vector<uint8_t>;
 using RegistrationSig = std::vector<uint8_t>;
 
 /// @brief A system generated part of the user's PRF key
-using S2 = std::vector<uint8_t>;
+// [TODO-UTT] Should be a byte vector similar to other opaque data types
+using S2 = std::vector<uint64_t>;
 
 /// @brief The privacy budget determines how much in value can a user transfer anonymously to other users.
 using PrivacyBudget = std::vector<uint8_t>;
